requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
boto3
botocore
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
dj-database-url==2.3.0
Django==4.2
django-environ==0.12.0
django-storages==1.14.5
djangorestframework==3.15.2
exceptiongroup==1.2.2
filelock==3.18.0
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
marshmallow==3.26.1
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.0.2
openai==1.68.2
orjson==3.10.15
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.7
propcache==0.3.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.0.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3
virtualenv==20.29.3
yarl==1.18.3
zstandard==0.23.0
